requirement.txt
Django==2.2
django-bootstrap-form==3.4
django-filter==2.2.0
django-simple-history==2.8.0
djangorestframework==3.11.0
mysqlclient==1.4.6

